relocate: Mypage 폴더 안에 있는 컴포넌트 폴더 정리
