+++ src/pages/Mypage/components/AddFavoriteIdolList/AddFavoriteIdolList.css
@@ -1,4 +1,4 @@
-@import "../assets/MyPageSlide.css";
+@import "../../assets/MyPageSlide.css";
 
 .add-favorite-idol-list-wrapper {
   margin-bottom: 48px;
