+++ src/pages/Mypage/components/AddFavoriteIdolList/AddIdolListButton/AddIdolListButton.css
