+++ src/pages/Mypage/components/AddFavoriteIdolList/AddIdolListItem/AddIdolListItem.css
