+++ src/pages/Mypage/components/MyFavoriteIdolList/MyFavoriteIdolList.css
