+++ src/pages/Mypage/components/MyFavoriteIdolList/MyIdolListItem/MyIdolListItem.css
