+++ src/pages/Mypage/components/MyFavoriteIdolList/MyIdolListItem/MyIdolListItem.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
-
 import './MyIdolListItem.css';
-import deleteButton from '../assets/images/btn_delete.svg';
+import deleteButton from '../../../assets/images/btn_delete.svg';
 
 function MyIdolListItem({ idolData, onClick }) {
   const { id, name, group, profilePicture } = idolData;
